Do not limit metalinks to repomd.xml
MirrorManager can track for multiple files the checksums and generate
the corresponding metalinks. In this implementation of the mirrorlist
server there was a check to ignore metalinks for anything not
'repomd.xml'. This removes the check to allow metalinks for all files if
the necessary data is available.

Also see: https://github.com/FedoraQt/MediaWriter/issues/78

Signed-off-by: Adrian Reber <adrian@lisas.de>