Add generate-mirrorlist-cache
generate-mirrorlist-cache generates the protobuf file needed to run
mirrorlist-server.

This is a re-implementation of the python script
mm2_refresh_mirrorlist_cache. The problem with the python version in
Fedora's infrastructure is that it needs between 35 and 50 minutes and
up to 10GB.

This Rust implementation uses about 600MB and takes only 1 minute. The
biggest difference is that the Python implementation let's the database
to all the work with SQL commands with a lot of joins. The Rust version
works with almost no joins and does all the complicated operations
locally.

Signed-off-by: Adrian Reber <adrian@lisas.de>