Only count mirrors which have the requested content
Before using global mirrors the mirrorlist servers looks for mirrors in
the same network, ASN, country and continent. If the number of mirrors
is higher than 'minimum' global mirrors are ignored. Before this change
we only checked if a mirror exists in the vicinity but not if it
actually carries the content the user is looking for. Now we actually
check if the mirror has the requested files.

Signed-off-by: Adrian Reber <adrian@lisas.de>